tests/wpt/tests/css/css-sizing/stretch/block-height-2.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#stretch-fit-sizing">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<meta name="assert" content="Checks the behaviour of stretch in various configurations.">

<body onload="checkLayout('[data-expected-height]')">

<div style="height: 200px; border: solid;">
  <div style="height: stretch; margin: 10px;" data-expected-height="180"></div>
</div>

<!--

Chrome fails the following test, which is a variation of example 9 in the spec. Safari passes.

From https://drafts.csswg.org/css-sizing-4/#stretch-fit-sizing, height:stretch for blocks behaves as
"""
100% but applying the resulting size to its margin box instead of the box indicated by box-sizing. For this purpose, auto margins are treated as zero, and furthermore, for block-level boxes in particular, if its block-start/block-end margin would be adjoining to its parent’s block-start/block-end margin if its parent’s sizing properties all had their initial values, then ***its block-start/block-end margin is treated as zero.***
"""

So the spec demands that the child has 0px margins and as a corollary, has 300px inner and outer height.

But Blink gives it an inner height of 200px and margins as specified (25px/75px).
-->
<div style="height: 300px; outline: 1px solid; margin: 10px 0px 10px 0px;">
  <div
    style="height: stretch; margin: 25px 0px 75px 0px; outline: 2px dashed blue;"
    data-expected-height="300"></div>
</div>
